Refactor RPCResponse into .Ok, .Error, .Completed
